pyproject.toml
[project]
name = "sr2-relay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Diego Camizotti", email = "diego@terminuslabs.co" }]
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "litellm[proxy]>=1.83.13",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.3",
    "uvicorn[standard]",
]

[project.scripts]
sr2-relay = "sr2_relay.cli:run"

[build-system]
requires = ["uv_build>=0.11.0,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = ["pytest>=9.0.3", "ruff>=0.15.11"]
